Port tests to STM32L0x2
This new test app doesn't use GDB/OpenOCD, like the other apps, instead
flashing the microcontroller using the built-in STM32 bootloader. The
reason for this is that the board I'm testing with doesn't have an
on-board programmer, and my ST-LINK is currently having problems (I've
ordered new ones but they haven't arrived yet).

This means that `panic-semihosting` doesn't work. I've implemented a
custom panic handler that lights an LED in case of panic.

Stage 1 and 2 from the README work. I observe no panics, and the test
suite in stage 2 runs successfully. I wasn't able to run stage 3. As far
as I can tell, the relevant tool on the linked page is only available on
Windows. I only have a Windows installation in a virtual machine, but I
wasn't able to get the testing tool to work in there.